Add user object to request after decoding token. (flask & falcon)
