+++ test/conftest.py
@@ -1,55 +1,74 @@
+import json
+from datetime import datetime, timedelta
+
+import jwt.algorithms
 import pytest
-import falcon
-from falcon import testing as falcon_testing
-from flask import Flask
 
-from thunderstorm_auth.testing import *  # noqa
-from thunderstorm_auth.falcon import TsAuthMiddleware
-from thunderstorm_auth.flask import ts_auth_required
-
-
-# FALCON FIXTURES
+from thunderstorm_auth import utils
 
 
 @pytest.fixture
-def resource():
-    class Resource:
-
-        requires_auth = True
-
-        def on_get(self, req, resp):
-            resp.body = 'ok'
-
-    return Resource()
+def private_key():
+    return utils.generate_private_key()
 
 
 @pytest.fixture
-def middleware(jwk_set):
-    return TsAuthMiddleware(jwk_set)
+def jwk(private_key):
+    jwk = jwt.algorithms.RSAAlgorithm.to_jwk(private_key[0].public_key())
+    jwk_dict = json.loads(jwk)
+    return {private_key[1]: jwk_dict}
 
 
 @pytest.fixture
-def falcon_app(resource, middleware):
-    app = falcon.API(middleware=middleware)
-    app.add_route('/', resource)
-    return app
+def alternate_private_key():
+    return utils.generate_private_key()
 
 
 @pytest.fixture
-def client(falcon_app):
-    return falcon_testing.TestClient(falcon_app)
+def alternate_jwk(alternate_private_key):
+    jwk = jwt.algorithms.RSAAlgorithm.to_jwk(alternate_private_key[0].public_key())
+    jwk_dict = json.loads(jwk)
+    return {alternate_private_key[1]: jwk_dict}
 
-
-# FLASK FIXTURES
 
 @pytest.fixture
-def flask_app(jwk_set):
-    app = Flask('test_app')
-    app.config['TS_AUTH_JWKS'] = jwk_set
+def jwk_set(jwk, alternate_jwk):
+    jwks = {}
+    # merge dictionaries
+    jwks.update(jwk)
+    jwks.update(alternate_jwk)
+    return {"keys": jwks}
 
-    @app.route('/')
-    @ts_auth_required
-    def hello_world():
-        return 'Hello, World!'
 
-    return app
+@pytest.fixture
+def token_data():
+    return {
+        'username': 'test-user',
+        'permissions': {},
+        'groups': []
+    }
+
+
+@pytest.fixture
+def valid_token(private_key, token_data):
+    return utils.encode_token(private_key, token_data)
+
+
+@pytest.fixture
+def invalid_token():
+    # using a junk string here rather than a truncated token as truncated tokens do not trigger the desired error
+    return 'this is not even a token'.encode('utf-8')
+
+
+@pytest.fixture
+def invalid_token_no_headers(private_key):
+    return jwt.encode({'data': 'nodata'}, private_key[0], algorithm='RS512')
+
+
+@pytest.fixture
+def expired_token(private_key, token_data):
+    expiry = datetime.utcnow() - timedelta(hours=1)
+    return utils.encode_token(
+        private_key,
+        dict(token_data, exp=expiry)
+    )
